% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Brathen_1982}
\alias{Brathen_1982}
\title{Brathen 1982 sapwood data set.}
\format{
A tibble of 2 variables:
\describe{
\item{n_sapwood}{number of sapwood rings}
\item{count}{number of times n_sapwood was observed}
}
}
\source{
Bräthen A. 1982. A tree-ring chronology from the western part of Sweden. Sapwood and a dating problem, in: Hackens T., Mejdahl V. (Eds.), Second Nordic Conference on the Application of Scientific Methods in Archaeology, PACT 7(1). pp. 27–35.
}
\usage{
Brathen_1982
}
\description{
Sapwood data set for Western Sweden published by Bräthen in 1982.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{sample size = 69 observations
}\if{html}{\out{</div>}}
}
\examples{
sw_data_info("Brathen_1982")

sw_model("Brathen_1982", plot = TRUE)

}
\keyword{datasets}
