% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_report.R
\name{fd_report}
\alias{fd_report}
\title{fd_report: a function the reports felling dates}
\usage{
fd_report(
  x,
  series = "series",
  last = "last",
  n_sapwood = "n_sapwood",
  waneyedge = "waneyedge",
  sw_data = "Hollstein_1980",
  credMass = 0.954,
  densfun = "lognormal",
  sep = ";"
)
}
\arguments{
\item{x}{Name of a \code{data.frame} with at least four columms, providing
information on the id's of the tree-ring series, the number of sapwood
rings observed, the presence of waney edge and the date assigned to the
last measured ring. A column describing the sapwood data set to be used
for modelling and the computation of the hdi can be provided as well.}

\item{series}{Name of the column in \code{x} where id's of the tree-ring series
are listed as \code{character} values.}

\item{last}{Name of the column in \code{x} which lists
the calendar year assigned to the last measured ring. Should be \code{numeric.}}

\item{n_sapwood}{Name of the column in \code{x} where the number of observed sapwood
rings are listed. This variable should be \code{numeric}.}

\item{waneyedge}{Name of the column in \code{x} indicating the presence
(\code{TRUE})/absence (\code{FALSE}) of waney edge. Should be a \code{logical} vector.}

\item{sw_data}{The name of the sapwood data set to use for modelling.
Should be one of \code{\link[=sw_data_overview]{sw_data_overview()}}, or the path to a .csv file with
columns ´n_sapwood´ and ´count´.}

\item{credMass}{A \code{scalar [0, 1]} specifying the mass within the credible
interval (default = .954).}

\item{densfun}{Name of the density function fitted to the sapwood data set.
Should be one of:
\itemize{
\item "lognormal" (the default value),
\item "normal",
\item "weibull",
\item "gammma".
}}

\item{sep}{Should be "," (comma)  or ";" (semi-colon) and is used when a
sapwood data set is provided from user-defined .csv-file.}
}
\value{
data.frame
}
\description{
This function reports the individual felling date of tree-ring series, based
on the presence/absence of sapwood and/or waney edge. This can take three
modes:
\itemize{
\item a \emph{terminus post quem} or \emph{earliest possible felling date}, when only
heartwood rings were recorded.
\item a felling date range or interval: when sapwood rings were recorded, but
no bark or waney edge is present.
\item an exact felling date: when bark or waney edge is present on the
measured sample.
}

Reports the lower and upper boundaries of a felling date range.
}
\examples{
tmp <- data.frame(id = c("aaa", "bbb", "ccc"),
                  swr = c(10, 11, 12),
                  waneyedge = c(FALSE, FALSE,TRUE),
                  end = c(10, 0, -10))
fd_report(tmp,
          series = "id",
          n_sapwood = "swr",
          last = "end",
          sw_data = "Wazny_1990")

}
