% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fellingdateR-package.R
\docType{package}
\name{fellingdateR-package}
\alias{fellingdateR}
\alias{fellingdateR-package}
\title{fellingdateR: `fellingdateR` is an R package that aims to facilitate the organisation, analysis and interpretation of tree-ring data from wooden heritage objects and structures}
\description{
`fellingdateR` is an R package that aims to facilitate the organization, analysis and interpretation of tree-ring data from wooden cultural heritage objects and structures.
}
\author{
\strong{Maintainer}: Kristof Haneca \email{Kristof.Haneca@vlaanderen.be} (\href{https://orcid.org/0000-0002-7719-8305}{ORCID})

Other contributors:
\itemize{
  \item Koen Van Daele (\href{https://orcid.org/0000-0002-8153-2978}{ORCID}) [contributor]
}

}
\keyword{internal}
