% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fellingdateR-package.R
\docType{package}
\name{fellingdater}
\alias{fellingdater-package}
\alias{fellingdater}
\title{fellingdater: Tree-ring dating and estimating felling dates of historical timbers}
\description{
fellingdater provides a comprehensive suite of functions
for dendrochronological and dendroarchaeological analysis, covering a
workflow from tree-ring \strong{data processing} and \strong{crossdating}, up to the
estimation and reporting of \strong{felling dates}. Originally designed to assist
dendrochronologists in inferring felling dates from dated tree-ring series
with partially preserved sapwood, the package has been expanded to include a
workflow for tree-ring dating, analysis, and visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/fellingdater}
  \item \url{https://github.com/ropensci/fellingdater}
  \item \url{https://ropensci.github.io/fellingdater}
  \item Report bugs at \url{https://github.com/ropensci/fellingdater/issues}
}

}
\author{
\strong{Maintainer}: Kristof Haneca \email{Kristof.Haneca@vlaanderen.be} (\href{https://orcid.org/0000-0002-7719-8305}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Koen Van Daele (\href{https://orcid.org/0000-0002-8153-2978}{ORCID}) [contributor]
  \item Ronald Visser (\href{https://orcid.org/0000-0001-6966-1729}{ORCID}) [contributor]
  \item Antonio Jesus Pérez-Luque (\href{https://orcid.org/0000-0002-1747-0469}{ORCID}) [reviewer]
  \item Nicholas John Tierney (\href{https://orcid.org/0000-0003-1460-8722}{ORCID}) [reviewer]
}

}
\keyword{package}
