% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hollstein_1980}
\alias{Hollstein_1980}
\title{Hollstein 1980 sapwood data set.}
\format{
A tibble of 2 variables:
\describe{
\item{n_sapwood}{number of sapwood rings}
\item{count}{number of times n_sapwood was observed}
}
}
\source{
Hollstein E. 1980. Mitteleuropäische Eichenchronologie: Trierer dendrochronologische Forschungen zur Archäologie und Kunstgeschichte, Trierer Grabungen und Forschungen. Verlag Phillipp von Zabern, Mainz am Rhein.
}
\usage{
Hollstein_1980
}
\description{
Sapwood data set for South and Central Germany published by Hollstein in 1980.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{sample size = 490 observations
}\if{html}{\out{</div>}}
}
\examples{
sw_data_overview()
sw_data_info("Hollstein_1980")
sw_model("Hollstein_1980")

}
\keyword{Hollstein_1980}
