% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Miles_1997_NM}
\alias{Miles_1997_NM}
\title{Miles 1997 sapwood data set.}
\format{
A tibble of 2 variables:
\describe{
\item{n_sapwood}{number of sapwood rings}
\item{count}{number of times n_sapwood was observed}
}
}
\source{
Miles D. 1997. The interpretation, presentation and use of tree-ring dates. Vernacular architecture 28, 40–56. DOI: https://doi.org/10.1179/030554797786050563.
}
\usage{
Miles_1997_NM
}
\description{
Sapwood data set for the Northern Midland counties (U.K.) - Cheshire, Staffordshire, West Midlands, Northamptonshire, Cambridgeshire, and everything to the north - published by Miles in 1997.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{sample size = 295 observations
}\if{html}{\out{</div>}}
}
\examples{
sw_data_overview()
sw_data_info("Miles_1997_NM")
sw_model("Miles_1997_NM")

}
\keyword{datasets}
