% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Miles_1997_SC}
\alias{Miles_1997_SC}
\title{Miles 1997 sapwood data set.}
\format{
A tibble of 2 variables:
\describe{
\item{n_sapwood}{number of sapwood rings}
\item{count}{number of times n_sapwood was observed}
}
}
\source{
Miles D. 1997. The interpretation, presentation and use of tree-ring dates. Vernacular architecture 28, 40–56. \url{https://doi.org/10.1179/030554797786050563}
}
\usage{
Miles_1997_SC
}
\description{
Sapwood data set for the Southern counties (U.K.), up to and including Gloucestershire, Warwickshire, Bedfordshire, Suffolk and Norfolk, published by Miles in 1997.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{sample size = 406 observations
}\if{html}{\out{</div>}}
}
\examples{
sw_data_info("Miles_1997_SC")

sw_model("Miles_1997_SC", plot = TRUE)

}
\keyword{datasets}
