% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Miles_1997_WBC}
\alias{Miles_1997_WBC}
\title{Miles 1997 sapwood data set.}
\format{
A tibble of 2 variables:
\describe{
\item{n_sapwood}{number of sapwood rings}
\item{count}{number of times n_sapwood was observed}
}
}
\source{
Miles D. 1997. The interpretation, presentation and use of tree-ring dates. Vernacular architecture 28, 40–56. DOI: https://doi.org/10.1179/030554797786050563.
}
\usage{
Miles_1997_WBC
}
\description{
Sapwood data set for Wales and border counties (U.K.), Shropshire, Hereford and Worcesterthe, published by Miles in 1997.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{sample size = 219 observations
}\if{html}{\out{</div>}}
}
\examples{
sw_data_overview()
sw_data_info("Miles_1997_WBC")
sw_model("Miles_1997_WBC")

}
\keyword{datasets}
