% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movAv.R
\name{movAv}
\alias{movAv}
\title{movAV: computes a running mean on a time series}
\usage{
movAv(x, w = 11, align = "center", edges = "fill")
}
\arguments{
\item{x}{A \code{numeric} vector of \code{length(x)}.}

\item{w}{The width of the window. When w is even, one more value from the
future is included.}

\item{align}{Should be "center (default), "left" or "right".
\itemize{
\item \code{align = "center"}: average assigned to the center of the window.
\item \code{align = "left"}: average of current value and next (w-1) values.
\item \code{align = "right"}: average of current value and previous (w-1) values
}}

\item{edges}{Defines how values are computed at the start and end of the
series:
\itemize{
\item \code{edges = "fill"}: average of reducing number of values near the edges of
the vector
(this is the default setting).
\item \code{edges = "nofill"}: fill with \code{NA} when window does not include w values.
}}
}
\value{
A \code{numeric} vector of \code{length(x)}.
}
\description{
This is a vectorized function that computes a running mean/moving average
over a window width \code{w} on a continuous (time) series.
}
