% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sohar_2012_ELL_t}
\alias{Sohar_2012_ELL_t}
\title{Sohar et al. 2012 sapwood data set.}
\format{
A tibble of 2 variables:
\describe{
\item{n_sapwood}{number of sapwood rings}
\item{count}{number of times n_sapwood was observed}
}
}
\source{
Sohar K., Vitas A. & Läänelaid A. 2012. Sapwood estimates of pedunculate oak (\emph{Quercus robur} L.) in eastern Baltic, Dendrochronologia 30.1, 49–56. \url{https://doi.org/10.1016/j.dendro.2011.08.001}
}
\usage{
Sohar_2012_ELL_t
}
\description{
Sapwood data set for Eastern Estonia, Latvia, Lithuania, published by Sohar et al. in 2012.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{sample size = 533 observations
!!! sapwood determined by presence of _tyloses_
}\if{html}{\out{</div>}}
}
\examples{
sw_data_info("Sohar_2012_ELL_t")

sw_model("Sohar_2012_ELL_t", plot = TRUE)

}
\keyword{datasets}
