% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sohar_2012_FWE_c}
\alias{Sohar_2012_FWE_c}
\title{Sohar et al. 2012 sapwood data set.}
\format{
A tibble of 2 variables:
\describe{
\item{n_sapwood}{number of sapwood rings}
\item{count}{number of times n_sapwood was observed}
}
}
\source{
Sohar K., Vitas A. & Läänelaid A. 2012. Sapwood estimates of pedunculate oak (\strong{Quercus robur} L.) in eastern Baltic, Dendrochronologia 30.1, 49–56. DOI: https://doi.org/10.1016/j.dendro.2011.08.001
}
\usage{
Sohar_2012_FWE_c
}
\description{
Sapwood data set for Southern Finland and western Estonia, published by Sohar et al. in 2012.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{sample size = 98 observations
!!! sapwood determined by *color*
}\if{html}{\out{</div>}}
}
\examples{
sw_data_overview()
sw_data_info("Sohar_2012_FWE_c")
sw_model("Sohar_2012_FWE_c")

}
\keyword{datasets}
