% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_combine.R
\name{sw_combine}
\alias{sw_combine}
\title{Compute a single felling date range by combining multiple sapwood estimates}
\usage{
sw_combine(
  x,
  series = "series",
  last = "last",
  n_sapwood = "n_sapwood",
  waneyedge = "waneyedge",
  sw_data = "Hollstein_1980",
  densfun = "lognormal",
  cred_mass = 0.954,
  hdi = TRUE,
  plot = FALSE
)
}
\arguments{
\item{x}{A \code{data.frame} with at least four columns, providing information on
the ID's of the tree-ring series, the number of sapwood rings observed, the
presence of waney edge and the calendar date assigned to the last measured
ring. A column describing the sapwood data set to be used for modelling and
the computation of the hdi can be provided as well.}

\item{series}{Name of the column in \code{x} where ID's of the tree-ring series
are listed as \code{character} values.}

\item{last}{Name of the column in \code{x} where calendar years assigned to the
last measured ring are listed (should be \code{numeric}).}

\item{n_sapwood}{Name of the column in \code{x} where the number of observed
sapwood rings are listed (should be \code{numeric}).}

\item{waneyedge}{Name of the column in \code{x} indicating the presence
(\code{TRUE})/absence (\code{FALSE}) of waney edge (should be a \code{logical} vector).}

\item{sw_data}{The name of the sapwood data set to use for modelling. It
should be one of the data sets listed in \code{\link[=sw_data_overview]{sw_data_overview()}}, or the name
of a \code{data.frame} with sapwood data in columns \code{n_sapwood} and \code{count}.}

\item{densfun}{Name of the density function fitted to the sapwood data set.
Should be one of:
\itemize{
\item \emph{lognormal} (the default value),
\item \emph{normal},
\item \emph{weibull},
\item \emph{gammma}.
}}

\item{cred_mass}{A \code{scalar [0, 1]} specifying the mass within the credible
interval (default = .954).}

\item{hdi}{A \code{logical} parameter. If \code{TRUE}, the lower and upper limit of the
highest density interval (credible interval) is given for the combined
felling date. When \code{FALSE}, a matrix is returned with scaled p values for
calendar years covering the combined estimate of the felling date range.}

\item{plot}{A \code{logical} parameter. If \code{TRUE} a ggplot-plot style graph is
returned of the individual and combined estimate of the felling date. If
\code{FALSE}, a list with numeric output of the modelling process is returned.}
}
\value{
Depends on the value of plot. If \code{TRUE} a ggplot-style is returned,
when \code{FALSE} a \code{data.frame} with the combined probability ($raw_data$comb),
the modelling parameters ($sapwood_data, $sapwood_model, $cred_mass),
the Highest Posterior Density Interval for combined combined felling date
interval ($hdi_combine), and the Agreement indices for the model ($A_model)
and individual series ($individual_series$agr_index).
}
\description{
This function assesses whether it is possible to combine
multiple sapwood estimates into a single felling date range.
}
\examples{
# a data set in which all series have preserved sapwood

sw_combine(trs_example1, plot = TRUE)

# a data set in which one series has an exact felling date (= waney edge preserved)

trs_example2
sw_combine(trs_example2, plot = TRUE)

# a data set in which multiples series have an exact felling date

trs_example3
sw_combine(trs_example3, plot = FALSE)

}
\seealso{
\code{\link[=sw_combine_plot]{sw_combine_plot()}}
}
