% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_combine_plot.R
\name{sw_combine_plot}
\alias{sw_combine_plot}
\title{Plot the output of \code{\link[=sw_combine]{sw_combine()}}}
\usage{
sw_combine_plot(x)
}
\arguments{
\item{x}{Output of \code{\link[=sw_combine]{sw_combine()}}.}
}
\value{
A ggplot-style graph, with calendar years on the X-axis and the
probability (p) on the Y-axis. Dots represent exact felling dates.
}
\description{
This function plots the result of \code{sw_combine()} and the
interval for, or the exact felling date of, a group of tree-ring series,
with agreement indices for the global model.
}
\examples{
sw_example2
tmp <- sw_combine(sw_example2)
sw_combine_plot(tmp)

}
