% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_data_overview.R
\name{sw_data_overview}
\alias{sw_data_overview}
\title{sw_data_overview: provides an overview all sapwood data sets that are
available within the \code{fellingdateR} package}
\usage{
sw_data_overview()
}
\value{
A \code{character} vector with the names of all available data sets
with sapwood counts.
}
\description{
This function provides an overview of available sapwood data
and models in the \code{fellingdateR} package.
}
\examples{
sw_data_overview()

}
