% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sw_example0}
\alias{sw_example0}
\title{Example dataset 0}
\format{
A tibble of 4 variables:
\describe{
\item{trs}{unique ID of the tree-ring series}
\item{end}{calendar year assigned to the last measured ring}
\item{swr}{number of observed sapwood rings}
\item{bark}{waney edge present TRUE/FALSE}
}
}
\usage{
sw_example0
}
\description{
A dataset in which all series have preserved sapwood.
Unconventional variable names.
}
\examples{
sw_example0

}
\keyword{datasets}
