% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sw_example2}
\alias{sw_example2}
\title{Example dataset 2}
\format{
A tibble of 4 variables:
\describe{
\item{series}{unique ID of the tree-ring series}
\item{last}{calendar year assigned to the last measured ring}
\item{n_sapwood}{number of observed sapwood rings}
\item{waneyedge}{waney edge present TRUE/FALSE}
}
}
\usage{
sw_example2
}
\description{
A dataset in which one series has an exact felling date (= waney edge present).
}
\examples{
sw_example2

}
\keyword{datasets}
