% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sw_example3}
\alias{sw_example3}
\title{Example dataset 3}
\format{
A tibble of 4 variables:
\describe{
\item{series}{unique ID of the tree-ring series}
\item{last}{calendar year assigned to the last measured ring}
\item{n_sapwood}{number of observed sapwood rings}
\item{waneyedge}{waney edge present TRUE/FALSE}
}
}
\usage{
sw_example3
}
\description{
A dataset with multiple exact felling dates.
}
\examples{
sw_example3

}
\keyword{datasets}
