% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_interval.R
\name{sw_interval}
\alias{sw_interval}
\title{Computes the limits of the felling date range}
\usage{
sw_interval(
  n_sapwood = NA,
  last = 1,
  hdi = FALSE,
  cred_mass = 0.954,
  sw_data = "Hollstein_1980",
  densfun = "lognormal",
  plot = FALSE
)
}
\arguments{
\item{n_sapwood}{A \code{numeric}. The number of observed sapwood rings.}

\item{last}{A \code{numeric}. The calendar year assigned to the outermost sapwood
ring (default = 1).}

\item{hdi}{Logical. If \code{TRUE}, the function returns the lower and upper bounds
of the highest density interval (HDI). If \code{FALSE}, it returns a data frame with
scaled probability values for each possible number of sapwood rings.}

\item{cred_mass}{A numeric \code{scalar} in the range of \verb{[0, 1]}, specifying the
mass of the credible interval (default = .954)}

\item{sw_data}{The name of the sapwood data set to use for modelling. It
should be one of the data sets listed in \code{\link[=sw_data_overview]{sw_data_overview()}}, or the name
of a \code{data.frame} with sapwood data in columns \code{n_sapwood} and \code{count}.}

\item{densfun}{A character string naming the distribution to fit.
One of:
\itemize{
\item \emph{lognormal} (default),
\item \emph{normal},
\item \emph{weibull},
\item \emph{gamma}.
}}

\item{plot}{A \code{logical}.
\itemize{
\item If \code{TRUE}, a ggplot-style graph is returned of the individual
sapwood model and estimate of the felling date range.
\item If \code{FALSE}, a \code{list} with the numeric output of the modelling process is returned.
}}
}
\value{
The type of output depends on the value of \code{hdi}:
\itemize{
\item If \code{hdi = TRUE}, a numeric vector with the lower and upper limits of the
highest density interval (HDI). Attributes include \code{cred_mass}, \code{sw_data},
and the fitted \code{model}.
\item If \code{hdi = FALSE}, a data frame with scaled probability values (\code{p}) for
each year after the last dated sapwood ring.
}
}
\description{
This function computes the probability density function (PDF)
and the highest posterior density interval (HDI) of the felling date range,
based on the observed number of sapwood rings, their dating, and a selected
sapwood dataset and model.
}
\examples{
# 10 sapwood rings observed and the Wazny 1990 sapwood model:
sw_interval(
     n_sapwood = 10,
     last = 1234,
     hdi = TRUE,
     cred_mass = .95,
     sw_data = "Wazny_1990",
     densfun = "lognormal",
     plot = FALSE
)
# same example as above, but with numerical output (hdi = FALSE):
sw_interval(
     n_sapwood = 10,
     last = 1234,
     hdi = FALSE,
     cred_mass = .95,
     sw_data = "Wazny_1990",
     densfun = "lognormal",
     plot = FALSE
)

}
