% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_sum.R
\name{sw_sum}
\alias{sw_sum}
\title{sw_sum: summed probabilities of felling date ranges.}
\usage{
sw_sum(
  x,
  series = "series",
  last = "last",
  n_sapwood = "n_sapwood",
  waneyedge = "waneyedge",
  sw_data = "Hollstein_1980",
  densfun = "lognormal",
  credMass = 0.954,
  plot = FALSE,
  scale_p = FALSE
)
}
\arguments{
\item{x}{data.frame containing columns c("series", "swr", "waneyedge" , "last"),
Column "sw_model" is optional. Could be the output of read_fh(x, header = TRUE)}

\item{series}{Name of the column containing the id's}

\item{last}{Name of the column containing the calendar year assigned to the last measured ring. Should be numeric.}

\item{n_sapwood}{Name of the column in \code{x} where the number of observed
sapwood rings are listed. This variable should be \code{numeric}.}

\item{waneyedge}{Name of the column indicating the presence (TRUE)/absence (FALSE) of waney edge. Should be logical}

\item{sw_data}{The name of the sapwood data set to use for modelling.
Should be one of \code{\link[=sw_data_overview]{sw_data_overview()}}, or the path to a .csv file with
columns ´n_sapwood´ and ´count´.}

\item{densfun}{Name of the density function fitted to the sapwood data set.
Should be one of:
\itemize{
\item "lognormal" (the default value),
\item "normal",
\item "weibull",
\item "gammma".
}}

\item{credMass}{A \code{scalar [0, 1]} specifying the mass within the credible
interval (default = .954).}

\item{plot}{logical, indicates whether a plot of the summed probability  distribution (SPD) should be returned.}

\item{scale_p}{logical, TRUE or FALSE.}
}
\value{
a data.frame
}
\description{
Computes the summed probability density (SPD) for a set of felling date ranges.
}
\examples{
dummy7 <- data.frame(
series = c("trs_40", "trs_41", "trs_42", "trs_43", "trs_44", "trs_45",
 "trs_46", "trs_47", "trs_48"),
last = c(1000, 1009, 1007, 1007, 1010, 1020, 1025, 1050, 1035),
n_sapwood = c(5, 10, 15, 16, 8, 0, 10, NA, 1),
waneyedge = c(FALSE, FALSE, TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, FALSE)
)

sw_sum(dummy7, plot = TRUE)

}
