% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trs_example5}
\alias{trs_example5}
\title{Example dataset 5}
\format{
A tibble of 4 variables:
\describe{
\item{series}{unique ID of the tree-ring series}
\item{last}{calendar year assigned to the last measured ring}
\item{n_sapwood}{number of observed sapwood rings}
\item{waneyedge}{waney edge present TRUE/FALSE}
}
}
\usage{
trs_example5
}
\description{
None of the series in this dataset have preserved sapwood rings.
}
\keyword{datasets}
