% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trs_zscore.R
\name{trs_zscore}
\alias{trs_zscore}
\title{Standardize tree-ring series to z-scores}
\usage{
trs_zscore(x)
}
\arguments{
\item{x}{A \code{data.frame} of class \code{"rwl"}, where each column is a tree-ring series.}
}
\value{
A \code{data.frame} of the same dimensions as \code{x}, with each column transformed to z-scores.
}
\description{
Converts each series (column) in an rwl-style \code{data.frame} to z-scores by subtracting the mean and dividing by the standard deviation.
}
\examples{
rwl <- trs_pseudo_rwl(n_series = 3, series_length = 50, end_date = 1990)
z_rwl <- trs_zscore(rwl)
apply(z_rwl, 2, mean, na.rm = TRUE) # should be ~0
apply(z_rwl, 2, sd, na.rm = TRUE) # should be ~1

}
