% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vanDaalen_Norway}
\alias{vanDaalen_Norway}
\title{van Daalen (unpublished) sapwood data set.}
\format{
A tibble of 2 variables:
\describe{
\item{n_sapwood}{number of sapwood rings}
\item{count}{number of times n_sapwood was observed}
}
}
\source{
S. van Daalen, unpublished dataset (version: 19 Dec 2022).
}
\usage{
vanDaalen_Norway
}
\description{
Sapwood data set for historical timbers found in the Netherlands that were
imported from Norway.
Unpublished data by S. van Daalen (version 19 Dec 2022, Van Daalen Dendrochronologie -
<www.dendro.nl>).
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{sample size = 104 observations
}\if{html}{\out{</div>}}
}
\examples{
sw_data_info("vanDaalen_Norway")

sw_model("vanDaalen_Norway")

}
\keyword{datasets}
