% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wazny_1990}
\alias{Wazny_1990}
\title{Wazny 1990 sapwood data set.}
\format{
A tibble of 2 variables:
\describe{
\item{n_sapwood}{number of sapwood rings}
\item{count}{number of times n_sapwood was observed}
}
}
\source{
Wazny T. 1990. Aufbau und Anwendung der Dendrochronologie für Eichenholz in Polen (PhD dissertation). Hamburg University, Hamburg.
}
\usage{
Wazny_1990
}
\description{
Sapwood data set for Poland, published by Wazny in 1990.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{sample size = 206 observations
}\if{html}{\out{</div>}}
}
\examples{
sw_data_overview()
sw_data_info("Wazny_1990")
sw_model("Wazny_1990")

}
\keyword{datasets}
