% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Weitz_2025}
\alias{Weitz_2025}
\title{Weitz 2025 sapwood data set.}
\format{
A tibble of 2 variables:
\describe{
\item{n_sapwood}{number of sapwood rings}
\item{count}{number of times n_sapwood was observed}
}
}
\source{
Weitz A., Haneca K. & Hoffsummer P. (in press): Estimating sapwood counts for historical oak timbers from Brussels, Belgium (13th-18th C), International Journal of Wood Culture.
}
\usage{
Weitz_2025
}
\description{
Sapwood data set for the Brussels region in Belgium, published by Weitz et al. in 2025.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{sample size = 443 observations
}\if{html}{\out{</div>}}
}
\examples{
sw_data_info("Weitz_2025")

sw_model("Weitz_2025", plot = TRUE)

}
\keyword{datasets}
