% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_table.R
\name{cor_table}
\alias{cor_table}
\title{cor_table: calculates correlation values between tree-ring series}
\usage{
cor_table(
  x,
  y = NULL,
  min_overlap = 50,
  remove.duplicates = TRUE,
  output = "table",
  values = c("r_pearson", "t_St", "glk", "glk_p", "t_BP", "t_Ho"),
  sort_by = "t_Ho"
)
}
\arguments{
\item{x}{Name of a \code{data.frame} (of class ´rwl´) with tree-ring data,
where each column represents a measurement series, and calendar years
are listed as rownames.}

\item{y}{Name of a \code{data.frame} (of class ´rwl´) with tree-ring data,
where each column represents a measurement series, and years are listed
as rownames.
x is compared to this series (y = reference).
If \code{NULL} (default) x is compared to itself (y = x).}

\item{min_overlap}{A \code{numeric} specifying the minimal overlap between series.}

\item{remove.duplicates}{Logical. If \code{TRUE} identical pairs of series and
references are removed from the output.}

\item{output}{Should be one of \code{matrix} or \code{table} (default).}

\item{values}{A character vector listing all correlation values that
should be computed. Defaults to all possible values:
c("r_pearson", "t_St", "glk", "glk_p", "t_BP", "t_Ho")}

\item{sort_by}{Correlation value by which the output is sorted for each
series in x.}
}
\value{
Depends on the value of \code{output}. If \code{matrix} a list of correlation
matrices is returned. If  \code{table} a data.frame is returned with all
correlation values listed by \code{values.}
}
\description{
This function calculates common correlation values between dated
tree-ring series (\code{x}) and a set of reference chronologies (\code{y}).
When no master chronologies are provided, each series in x is compared to all
other series in x.
}
