% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_report.R
\name{fd_report}
\alias{fd_report}
\title{Report felling dates of individual tree-ring series}
\usage{
fd_report(
  x,
  series = "series",
  last = "last",
  n_sapwood = "n_sapwood",
  waneyedge = "waneyedge",
  sw_data = "Hollstein_1980",
  cred_mass = 0.954,
  densfun = "lognormal"
)
}
\arguments{
\item{x}{Name of a \code{data.frame} with at least four columms, providing
information on
\itemize{
\item the id's of the tree-ring series ("series")
\item the number of sapwood rings observed ("n_sapwood)
\item the presence of waney edge ("waneyedge")
\item the year assigned to the last measured ring ("last").
}

Optionally, a column specifying the sapwood data set (\code{sw_data}) can also be included.}

\item{series}{Name of the column in \code{x} where id's of the tree-ring series
are listed as \code{character} values.}

\item{last}{Name of the column in \code{x} which lists the calendar year assigned
to the last measured ring (should be a \code{numeric} vector).}

\item{n_sapwood}{Name of the column in \code{x} where the number of observed
sapwood rings are listed (should be \code{numeric} vector).}

\item{waneyedge}{Name of the column in \code{x} indicating the presence
(\code{TRUE})/absence (\code{FALSE}) of waney edge (should be a \code{logical} vector).}

\item{sw_data}{There are two options:
\itemize{
\item A \code{character} string providing the name of the sapwood data set to use
for modelling. It should be one of the data sets listed in
\code{\link[=sw_data_overview]{sw_data_overview()}},
\item or the name of a \code{data.frame} with sapwood data
in columns \code{n_sapwood} and \code{count}, or
\item or character string naming a column in \code{x} that lists for each series
the sapwood model to use, e.g. sw_data = "sapwood_model_column".
}}

\item{cred_mass}{A numeric \code{scalar [0, 1]} specifying the mass within the
credible interval (default = .954).}

\item{densfun}{Name of the density function to fit to the sapwood
distribution. Should be one of:
\itemize{
\item \emph{lognormal} (the default value),
\item \emph{normal},
\item \emph{weibull},
\item \emph{gamma}.
}}
}
\value{
A \code{data.frame} with felling date estimates per tree-ring series.
Columns include:
\itemize{
\item \code{series}: series identifier
\item \code{last}: last measured ring
\item \code{n_sapwood}: number of sapwood rings
\item \code{waneyedge}: TRUE/FALSE for waney edge
\item \code{lower}, \code{upper}: numeric bounds for the estimated felling date
\item \code{felling_date}: a character summary (e.g., "between 1500 and 1510")
\item \code{sapwood_model}: the sapwood data/model used
}
}
\description{
This function reports the felling date estimate of individual
tree-ring series, based on the presence/absence of sapwood and/or waney
edge. There are three possible modes of reporting:
\itemize{
\item a \emph{terminus post quem} or \emph{earliest possible felling date}: when only
heartwood rings have been observed and measured
\item a felling date range or interval: when sapwood rings have been recorded, but
no bark or waney edge is present.
\item an exact felling date: when bark or waney edge is present on the
measured sample.
}
}
\examples{
df <- data.frame(
     id = c("trs1", "trs2", "trs3", "trs4"),
     swr = c(7, 1, 10, 12),
     waneyedge = c(FALSE, FALSE, FALSE, TRUE),
     end = c(1482, 1475, 1490, 1498)
)
fd_report(df,
     series = "id",
     n_sapwood = "swr",
     last = "end",
     sw_data = "Wazny_1990"
)

# Example with different sw_model for individual series
# You can add a user-defined sapwood dataset as well.

sapwood_model_column <- c(
     "Sohar_2012_ELL_c",
     "Wazny_1990",
     "Hollstein_1980",
     "vanDaalen_Norway"
)

df2 <- cbind(df, sw_data = sapwood_model_column)

fd_report(df2,
     series = "id",
     n_sapwood = "swr",
     last = "end",
     sw_data = "sw_data"
)

}
\seealso{
\code{\link[=sw_interval]{sw_interval()}}, \code{\link[=sw_data_overview]{sw_data_overview()}}, \code{\link[=sw_interval_plot]{sw_interval_plot()}}
}
