% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fellingdater-package.R
\docType{package}
\name{fellingdater}
\alias{fellingdater}
\alias{fellingdater-package}
\title{fellingdater}
\description{
fellingdater is an R package that facilitate the analysis and
interpretation of tree-ring data from wooden cultural heritage objects and
structures. The package standardizes the process of computing and combining
felling date estimates (fd), derived from the observed number of sapwood
rings (sw), both for individual and groups of related tree-ring series.
}
\section{Learn more about fellingdater at}{

\url{https://hanecakr.github.io/fellingdater/index.html/}
}

\seealso{
It's main functions are:
\itemize{
\item \code{\link[=sw_model]{sw_model()}}
\item \code{\link[=sw_interval]{sw_interval()}}
\item \code{\link[=fd_report]{fd_report()}}
\item \code{\link[=sw_combine]{sw_combine()}}
\item \code{\link[=sw_sum]{sw_sum()}}
}
}
\author{
\strong{Maintainer}: Kristof Haneca \email{Kristof.Haneca@vlaanderen.be} (\href{https://orcid.org/0000-0002-7719-8305}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Koen Van Daele (\href{https://orcid.org/0000-0002-8153-2978}{ORCID}) [contributor]
  \item Ronald Visser (\href{https://orcid.org/0000-0001-6966-1729}{ORCID}) [contributor]
  \item Antonio Jesus Pérez-Luque (\href{https://orcid.org/0000-0002-1747-0469}{ORCID}) [reviewer]
  \item Nicholas John Tierney (\href{https://orcid.org/0000-0003-1460-8722}{ORCID}) [reviewer]
}

}
\keyword{package}
