% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_header.R
\name{get_header}
\alias{get_header}
\title{get_header: retrieves the attributes of a .fh/Heidelberg format file}
\usage{
get_header(rwl)
}
\arguments{
\item{rwl}{The output of \code{read_fh(x, header = TRUE)},
a \code{data.frame} of class \code{rwl}.}
}
\value{
A \code{data.frame} with 26 header fields.
}
\description{
This function reports the HEADER fields from a .fh/Heidelberg
format file. The header fields are harvested from the .fh file by the
\code{read_fh()} function, which stores the HEADER fields in the .fh file as
attributes of the \code{data.frame} it returns.
}
