% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fh.R
\name{read_fh}
\alias{read_fh}
\title{read_fh: read a .fh/Heidelberg format file}
\usage{
read_fh(fname, BC_correction = FALSE, verbose = TRUE, header = FALSE)
}
\arguments{
\item{fname}{A \code{character} vector giving the name of the file name of
the fh file.}

\item{BC_correction}{A \code{logical}. If \code{TRUE} the correction moves BC-dates
one year forward.}

\item{verbose}{A \code{logical}. \code{FALSE} suppresses messages from being printed.}

\item{header}{A \code{logical}. if \code{TRUE} the HEADER fields are returned as
a \code{data.frame}, if \code{FALSE} the measurement data are returned.}
}
\value{
a \code{data.frame} containing ring-width measurements with calendar
years as \code{row.names} and header fields as \code{attributes.}
}
\description{
This function reads in a .fh/Heidelberg format file of ring widths AND lists
HEADER fields as attributes.
}
\details{
This reads in a fh-file with ring widths in blocks (decadal format)
or in columns (e.g., as with comment flags) as used by TSAP program.
Chronologies or half-chronos in fh-format are ALSO supported. The
\code{read_fh()} function is case insensitive. Information found in the HEADER
fields are listed as attributes.
}
\references{
This function is an extension of the \code{read.fh()} function from
the \strong{dplR package}, developped by Andy Bunn (Bunn 2008, Bunn 2010,
Bunn et al. 2022).
}
\author{
The original \code{read.fh()} function is part of the \strong{dplR package}
and was developped by Christian Zang, with new features and patches by Mikko
Korpela and Ronald Visser. This \code{read_fh()} function expands the
functionalities of the original \code{\link[dplR:read.fh]{dplR::read.fh()}} function.
}
