% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_combine.R
\name{sw_combine}
\alias{sw_combine}
\title{Combine multiple sapwood estimates into a single felling date range}
\usage{
sw_combine(
  x,
  series = "series",
  last = "last",
  n_sapwood = "n_sapwood",
  waneyedge = "waneyedge",
  sw_data = "Hollstein_1980",
  densfun = "lognormal",
  cred_mass = 0.954,
  hdi = TRUE,
  plot = FALSE
)
}
\arguments{
\item{x}{A \code{data.frame} with at least four columns, providing information on
the ID's of the tree-ring series, the number of sapwood rings observed, the
presence of waney edge and the calendar date assigned to the last measured
ring. A column describing the sapwood data set to be used for modelling and
the computation of the hdi can be provided as well.}

\item{series}{Name of the column in \code{x} where ID's of the tree-ring series
are listed as \code{character} values.}

\item{last}{Name of the column in \code{x} where calendar years assigned to the
last measured ring are listed (should be \code{numeric}).}

\item{n_sapwood}{Name of the column in \code{x} where the number of observed
sapwood rings are listed (should be \code{numeric}).}

\item{waneyedge}{Name of the column in \code{x} indicating the presence
(\code{TRUE})/absence (\code{FALSE}) of waney edge (should be a \code{logical} vector).}

\item{sw_data}{The name of the sapwood data set to use for modelling. It
should be one of the data sets listed in \code{\link[=sw_data_overview]{sw_data_overview()}}, or the name
of a \code{data.frame} with sapwood data in columns \code{n_sapwood} and \code{count}.}

\item{densfun}{Name of the density function fitted to the sapwood data set.
Should be one of:
\itemize{
\item \emph{lognormal} (the default value),
\item \emph{normal},
\item \emph{weibull},
\item \emph{gamma}.
}}

\item{cred_mass}{A \code{scalar [0, 1]} specifying the mass within the credible
interval (default = .954).}

\item{hdi}{A \code{logical} parameter. If \code{TRUE}, the lower and upper limit of the
highest density interval (credible interval) is given for the combined
felling date. When \code{FALSE}, a matrix is returned with scaled p values for
calendar years covering the combined estimate of the felling date range.}

\item{plot}{A \code{logical} parameter. If \code{TRUE} a ggplot-plot style graph is
returned of the individual and combined estimate of the felling date. If
\code{FALSE}, a list with numeric output of the modelling process is returned.}
}
\value{
@return Depends on \code{plot}:
\itemize{
\item If \code{TRUE}: a \code{ggplot}-style plot showing individual and combined felling date estimates.
\item If \code{FALSE}: a \code{list} with:
- \verb{$raw_data}: matrix of combined and individual densities.
- \verb{$sapwood_data}: sapwood data set used.
- \verb{$sapwood_model}: density function used.
- \verb{$cred_mass}: credibility mass for the interval.
- \verb{$hdi_combine}: HDI for the combined felling date range.
- \verb{$individual_series}: summary with agreement indices.
- \verb{$A_model}: overall model agreement index.
- \verb{$A_c}: critical threshold (default 60\%).
- \verb{$model_summary}: text summary of the result.Depends on the value of plot.
The rationale of the agreement index is outlined by Bronk Ramsey (1995; 2009)
}
}
\description{
This function evaluates whether multiple sapwood estimates can
be combined into a single felling date range and computes the resulting
interval if possible.
}
\examples{
# Example with all series preserving sapwood

sw_combine(sw_example1, plot = TRUE)

# Example with one exact felling date (= waney edge preserved)

sw_example2
sw_combine(sw_example2, plot = TRUE)

# Example with multiple exact felling dates

sw_example3
sw_combine(sw_example3, plot = FALSE)

}
\references{
\itemize{
\item Bronk Ramsey, C. (1995) Radiocarbon calibration and analysis of
stratigraphy: the OxCal program. \emph{Radiocarbon} \strong{37}, 425–430.
\url{https://doi.org/10.1017/S0033822200030903}
\item Bronk Ramsey, C. (2009) Bayesian analysis of radiocarbon dates.
\emph{Radiocarbon} \strong{51}, 337–360.
\url{https://doi.org/10.1017/S0033822200033865}
}
}
\seealso{
\code{\link[=sw_combine_plot]{sw_combine_plot()}}
}
