% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_data_info.R
\name{sw_data_info}
\alias{sw_data_info}
\title{sw_data_info: returns more detailed information on a sapwood data set}
\usage{
sw_data_info(x = NULL)
}
\arguments{
\item{x}{The name of a sapwood data set (see \code{\link[=sw_data_overview]{sw_data_overview()}} for
an overview).}
}
\value{
A \code{list} with:
\itemize{
\item \code{data}: the name of the data set,
\item \code{citation}: the correct citation for the data set,
\item \code{area}: the area represented by the data set,
\item \code{n_observations}: the no. of observations in the data set and,
\item \code{summary_raw_data}: summary stats of the raw data.
}
}
\description{
This function returns more information on a sapwdood data set. It lists the
correct citation for the data set, the area the data set is supposed to
cover, the number of observations and summary stats of the raw data.
}
\examples{
sw_data_info("Hollstein_1980")

sw_data_info("Sohar_2012_FWE_c")$citation

}
