% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_interval_plot.R
\name{sw_interval_plot}
\alias{sw_interval_plot}
\title{Plot function for the output of \code{\link[=sw_interval]{sw_interval()}}}
\usage{
sw_interval_plot(x)
}
\arguments{
\item{x}{Output of \code{\link[=sw_interval]{sw_interval()}}.}
}
\value{
A ggplot-style graph.
}
\description{
Returns a ggplot-style graph of the probability density function for the
felling date range, as computed by \code{\link[=sw_interval]{sw_interval()}}.
}
\examples{
tmp <- sw_interval(n_sapwood = 10,
            last = 1000,
           hdi = FALSE,
            credMass = .95,
            sw_data = "Hollstein_1980",
            densfun = "lognormal",
            plot = FALSE)
sw_interval_plot(tmp)

}
