% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_interval_plot.R
\name{sw_interval_plot}
\alias{sw_interval_plot}
\title{sw_interval_plot}
\usage{
sw_interval_plot(x, credMass = 0.954)
}
\arguments{
\item{x}{Output of \code{sw_interval()}.}

\item{credMass}{A \code{scalar} \verb{[0, 1]} specifying the mass within the credible
interval (default = .954).}
}
\value{
ggplot style graph.
}
\description{
sw_interval_plot
}
