% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_model.R
\name{sw_model}
\alias{sw_model}
\title{sw_model: modelling of sapwood data}
\usage{
sw_model(
  sw_data = "Hollstein_1980",
  densfun = "lognormal",
  credMass = 0.954,
  sep = ";",
  plot = TRUE
)
}
\arguments{
\item{sw_data}{The name of the sapwood data set to use for modelling.
Should be one of \code{\link[=sw_data_overview]{sw_data_overview()}}, or the path to a .csv file with
columns \code{n_sapwood} and \code{count}.}

\item{densfun}{Name of the density function fitted to the sapwood data set.
Should be one of:
\itemize{
\item "lognormal" (the default value),
\item "normal",
\item "weibull",
\item "gamma".
}}

\item{credMass}{A \code{scalar [0, 1]} specifying the mass within the credible
interval (default = .954).}

\item{sep}{Should be "," (comma)  or ";" (semi-colon) and is used when a
sapwood data set is provided from user-defined .csv-file.}

\item{plot}{\code{Logical.} If \code{TRUE} a plot is returned of the fitted density
function. When \code{FALSE} a list with numeric output of the modelling process
is returned.}
}
\value{
Depends on the \code{plot} parameter.
\itemize{
\item \code{plot = TRUE}: a ggplot-style graph.
\item \code{plot = FALSE}: a list with the numeric output of the modelling process.
}
}
\description{
This function fits a distribution to a data set of observed sapwood numbers
and computes the highest posterior density interval (hdi) for a given
credibility mass.
}
