% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_model.R
\name{sw_model}
\alias{sw_model}
\title{Model sapwood data and compute the highest posterior density interval}
\usage{
sw_model(
  sw_data = "Hollstein_1980",
  densfun = "lognormal",
  cred_mass = 0.954,
  source = NA,
  plot = FALSE
)
}
\arguments{
\item{sw_data}{The name of the sapwood data set to use for modelling. It
should be one of the data sets listed in \code{\link[=sw_data_overview]{sw_data_overview()}}, or the name
of a \code{data.frame} with sapwood data in columns \code{n_sapwood} and \code{count}.}

\item{densfun}{Name of the density function to fit to the sapwood data set.
Should be one of:
\itemize{
\item \emph{lognormal} (the default value),
\item \emph{normal},
\item \emph{weibull},
\item \emph{gammma}.
}}

\item{cred_mass}{A \code{scalar} in the range of \verb{[0, 1]} specifying the mass
within the credible interval (default = .954).}

\item{source}{A character string with info on the data source. This info is
incorporated in the output when a sapwood data set is provided from a
user-defined \code{data.frame}.}

\item{plot}{A \code{logical}. If \code{TRUE} a plot of the fitted density function is
returned. When \code{FALSE}, a list with numeric output of the modelling process
is returned.}
}
\value{
The return value depends on the \code{plot} parameter.
\itemize{
\item if \code{plot} is TRUE, a ggplot-style graph is returned.
\item if \code{plot} is FALSE, a list with the numeric output of the modelling
process is returned.
}
}
\description{
This function fits a distribution to a data set of observed
sapwood numbers and computes the highest posterior density interval (hdi)
for a given credibility mass.
}
\examples{
sw_model("Sohar_2012_ELL_c")

sw_model("Hollstein_1980")

}
