% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_sum.R
\name{sw_sum}
\alias{sw_sum}
\title{Compute the summed probability of multiple felling date ranges}
\usage{
sw_sum(
  x,
  series = "series",
  last = "last",
  n_sapwood = "n_sapwood",
  waneyedge = "waneyedge",
  sw_data = "Hollstein_1980",
  densfun = "lognormal",
  cred_mass = 0.954,
  plot = FALSE,
  scale_p = FALSE
)
}
\arguments{
\item{x}{A \code{data.frame} with columns c("series", "swr", "waneyedge" ,
"last"). Could be the output of \code{\link[=read_fh]{read_fh()}}.
Column "sw_model" is optional in order to specify the sapwood model to
use for each series.}

\item{series}{Name of the column in \code{x} where ID's of the tree-ring series
are listed as \code{character} values.}

\item{last}{Name of the column in \code{x} where calendar years assigned to the
last measured ring are listed (should be \code{numeric}).}

\item{n_sapwood}{Name of the column in \code{x} where the number of observed
sapwood rings are listed (should be \code{numeric})}

\item{waneyedge}{Name of the column in \code{x} indicating the presence
(\code{TRUE})/absence (\code{FALSE}) of waney edge (should be a \code{logical}).}

\item{sw_data}{The name of the sapwood data set to use for modelling. It
should be one of the data sets listed in \code{\link[=sw_data_overview]{sw_data_overview()}}, or the name
of a \code{data.frame} with sapwood data in columns \code{n_sapwood} and \code{count}.}

\item{densfun}{Name of the density function fitted to the sapwood data set.
Should be one of:
\itemize{
\item \emph{lognormal} (the default value),
\item \emph{normal},
\item \emph{weibull},
\item \emph{gamma}.
}}

\item{cred_mass}{A \code{scalar [0, 1]} specifying the mass within the credible
interval (default = .954).}

\item{plot}{A \code{logical}.
\itemize{
\item If \code{TRUE}, \code{\link[=sw_sum_plot]{sw_sum_plot()}} is triggered and a ggplot-style graph is
returned with the summed probability density (SPD).
\item If \code{FALSE}, a \code{list} with the numeric output of the modelling process is
returned.
}}

\item{scale_p}{A \code{logical}. If \code{TRUE} the summed probability density is
scaled to 1 (default).}
}
\value{
Depends on the value of \code{plot.}
}
\description{
Computes the summed probability density (SPD) for a set of
felling date ranges.
}
\examples{
sw_example7

sw_sum(sw_example7, densfun = "lognormal", cred_mass = 0.63, plot = FALSE)

}
\seealso{
\code{\link[=sw_sum_plot]{sw_sum_plot()}}
}
