% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_sum.R
\name{sw_sum}
\alias{sw_sum}
\title{Compute the summed probability of multiple felling date ranges}
\usage{
sw_sum(
  x,
  series = "series",
  last = "last",
  n_sapwood = "n_sapwood",
  waneyedge = "waneyedge",
  sw_data = "Hollstein_1980",
  densfun = "lognormal",
  credMass = 0.954,
  plot = FALSE,
  scale_p = FALSE,
  sep = ";"
)
}
\arguments{
\item{x}{A \code{data.frame} containing columns c("series", "swr", "waneyedge" ,
"last"), Column "sw_model" is optional. Could be the output of read_fh(x,
header = TRUE)}

\item{series}{Name of the column in \code{x} containing the ID's of the series.}

\item{last}{Name of the column in \code{x}containing the calendar year assigned to
the last measured ring (should be \code{numeric}).}

\item{n_sapwood}{Name of the column in \code{x} where the number of observed
sapwood rings are listed (should be \code{numeric}).}

\item{waneyedge}{Name of the column in \code{x} indicating the presence
(TRUE)/absence (FALSE) of waney edge (should be a \code{logical} vector).}

\item{sw_data}{The name of the sapwood data set to use for modelling. Should
be one of \code{\link[=sw_data_overview]{sw_data_overview()}}, or the path to a .csv file with columns
´n_sapwood´ and ´count´.}

\item{densfun}{Name of the density function fitted to the sapwood data set.
Should be one of:
\itemize{
\item "lognormal" (the default value),
\item "normal",
\item "weibull",
\item "gammma".
}}

\item{credMass}{A \code{scalar [0, 1]} specifying the mass within the credible
interval (default = .954).}

\item{plot}{A \code{logical}.
\itemize{
\item If \code{TRUE}, \code{\link[=sw_sum_plot]{sw_sum_plot()}} is triggered and a ggplot-style graph is
returned with the summed probability density (SPD).
\item If \code{FALSE}, a list with the numeric output of the modelling process is returned.
}}

\item{scale_p}{A \code{logical}. If \code{TRUE} the summed probability density is
scaled to 1.}

\item{sep}{Should be "," (comma)  or ";" (semi-colon) and is used when a
sapwood data set is provided from user-defined .csv-file.}
}
\value{
Depends on the value of \code{plot.}
}
\description{
Computes the summed probability density (SPD) for a set of
felling date ranges.
}
\examples{
dummy7 <- data.frame(
series = c("trs_40", "trs_41", "trs_42", "trs_43", "trs_44", "trs_45",
 "trs_46", "trs_47", "trs_48"),
last = c(1000, 1009, 1007, 1007, 1010, 1020, 1025, 1050, 1035),
n_sapwood = c(5, 10, 15, 16, 8, 0, 10, NA, 1),
waneyedge = c(FALSE, FALSE, TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, FALSE)
)

sw_sum(dummy7, plot = TRUE)

}
\seealso{
\code{\link[=sw_sum_plot]{sw_sum_plot()}}
}
