% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_sum_plot.R
\name{sw_sum_plot}
\alias{sw_sum_plot}
\title{Plot the output of \code{\link[=sw_sum]{sw_sum()}}}
\usage{
sw_sum_plot(
  x,
  bar_col = "steelblue",
  spline_col = "red3",
  dot_col = "steelblue4",
  dot_size = 2,
  dot_shape = 21
)
}
\arguments{
\item{x}{The output of \code{sw_sum()}.}

\item{bar_col}{The fill color for the bars.}

\item{spline_col}{The line color of the fitted smoothing spline.}

\item{dot_col}{The color of the shapes that represent exact felling dates.}

\item{dot_size}{A size argument for the shapes that represent exact felling dates.}

\item{dot_shape}{Number corresponding to the point symbol available in R for the shapes that represent exact felling dates.}
}
\value{
A ggplot style graph.
}
\description{
This function plots the result of \code{sw_sum()} - both the SPD and
the occurrence of exact felling dates - and adds a smoothing spline to the
SPD.
}
\examples{
trs_example6 <- trs_example6
trs_example6
tmp <- sw_sum(trs_example6)
sw_sum_plot(tmp, bar_col = "burlywood1")

}
