% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_sum_plot.r
\name{sw_sum_plot}
\alias{sw_sum_plot}
\title{sw_sum_plot: plot the output of \code{sw_sum()}}
\usage{
sw_sum_plot(x, bar_col = "lightblue", spline_col = "tomato3")
}
\arguments{
\item{x}{The output of \code{sw_sum()}.}

\item{bar_col}{The fill color for the bars.}

\item{spline_col}{The line color of the fitted smoothing spline.}
}
\value{
A ggplot style graph.
}
\description{
This function plots the result of \code{sw_sum()} - both the SPD and
the occurrence of exact felling dates - and adds a smoothing spline to the SPD.
}
\examples{
dummy6 <- fellingdateR:::dummy6
dummy6
tmp <- sw_sum(dummy6)
sw_sum_plot(tmp, bar_col = "burlywood1")

}
