% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_sum_plot.r
\name{sw_sum_plot}
\alias{sw_sum_plot}
\title{sw_sum_plot: plot the output of \code{sw_sum()}}
\usage{
sw_sum_plot(x, bar_col = "lightblue", spline_col = "tomato3")
}
\arguments{
\item{x}{The output of \code{sw_sum()}.}

\item{bar_col}{The fill color for the bars.}

\item{spline_col}{The line color of the fitted smoothing spline.}
}
\value{
A ggplot style graph.
}
\description{
This function plots the result of \code{sw_sum()} - both the SPD and
the occurrence of exact felling dates - and adds a smoothing spline to the SPD.
}
\examples{
dummy6 <- data.frame(
series = c("trs_30", "trs_31", "trs_32", "trs_34", "trs_35", "trs_36",
"trs_37", "trs_38", "trs_39"),
last = c(1000, 1009, 1007, 1005, 1010, 1020, 1025, 1050, 1035),
n_sapwood = c(5, 10, 15, 16, 8, 0, 10, 3, 1),
waneyedge = c(FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE)
)
tmp <- sw_sum(dummy6)
sw_sum_plot(tmp, bar_col = "burlywood1")

}
