% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trs_tbp_transform.R
\name{trs_tbp_transform}
\alias{trs_tbp_transform}
\title{Pre-transform tree-ring series using Baillie & Pilcher method}
\usage{
trs_tbp_transform(x)
}
\arguments{
\item{x}{A rwl-style data frame.}
}
\value{
A matrix of log-transformed index values.
}
\description{
Applies a 5-year centered moving average and computes the log index:
\code{log(100 * value / MA5)}. Uses edge mirroring to maintain
constant window size and avoid NA values.
}
\examples{
# Create sample data,
trs <- trs_pseudo_rwl(n_series = 5, series_length = 100, end_date = c(2020, 2025))
transformed <- trs_tbp_transform(trs)
head(transformed, 10)

}
